Update clang-tidy suggestions

+++ src/fheroes2/heroes/heroes.cpp
@@ -991,7 +991,7 @@ bool Heroes::PickupArtifact( const Artifact & art )
     if ( isControlHuman() ) {
         for ( const ArtifactSetData & artifactSetData : assembledArtifacts ) {
             const fheroes2::ArtifactDialogElement artifactUI( artifactSetData._assembledArtifactID );
-            fheroes2::showMessage( fheroes2::Text( Artifact( artifactSetData._assembledArtifactID ).GetName(), fheroes2::FontType::normalYellow() ),
+            fheroes2::showMessage( fheroes2::Text( Artifact( static_cast<int>( artifactSetData._assembledArtifactID ) ).GetName(), fheroes2::FontType::normalYellow() ),
                                    fheroes2::Text( _( artifactSetData._assembleMessage ), fheroes2::FontType::normalWhite() ), Dialog::OK, { &artifactUI } );
         }
     }
